riscv: Add hart feature run-time detection framework
This patch introduces a framework to detect and store hart features
(e.g. ISA extensions and their parameters) for RISC-V.

This patch does not introduce a concrete mechanism for run-time
detection, but implements everything so that such a mechanism
can be introduced.

Most of the changes in this patch are inspired by similar code
for other architectures, so nothing surprising should be hidden here.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>