RISC-V: Move TSO check to elf_machine_matches_host()
The current implementation does not support ELF files which  have the
EF_RISCV_TSO ELF flag set.  This is implemented in process_elf64_file(),
with the assumption that no TSO-capable hosts exists.

In order to weaken this restriction, let's move the TSO flag check
into elf_machine_matches_host(), where compatibility with the host
can be checked.

This patch serves as a preparation for further changes and does
not intend to implement any functional changes.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>