RISC-V: Attempt to enable TSO mode for TSO binaries
The upcoming RISC-V Ssdtso specification introduces a bit in the senvcfg
CSR to switch the memory consistency model at run-time from RVWMO to TSO
(and back).  The active consistency model can therefore be switched on a
per-hart base and managed by the kernel on a per-process/thread base.

A RFC kernel patchset has been posted that provides a prctl API to get
and set the current consistency model.

This patch attempts to switch to the TSO consistency model in case
the ELF file requires a TSO machine and the machine does not run
in TSO mode.  If the attempt fails, we fall back to the old behaviour
and claim that the ELF file is not compatible with the host.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>